requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
crispy-tailwind==1.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
